tgbot/requirements.txt
aiocron==1.3
aiogram==2.9.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
croniter==1.0.1
idna==2.10
loguru==0.5.3
motor==2.1.0
multidict==4.7.6
natsort==7.1.0
Pillow==7.2.0
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
tzlocal==2.1
ujson==3.1.0
yarl==1.5.1
